+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ### Added
 
-- CLI: `only_if_older` option to refresh to skip refreshing if recently refreshed
+- CLI: add `--only-if-older=<time>` to `refresh` sub command to skip if recently refresed
+- api: `only_if_older` option to refresh to skip refreshing if recently refreshed
 
 ## 0.4.0